# Don't re-check the secret file on disk multiple times per request