Don't re-check the secret file on disk multiple times per request
    
This is a performance improvement and consistency change. For the
lifetime of a single server span, we use the same cached-in-memory copy
of secrets rather than re-checking disk every time and only fetch at all
if we check the secrets. This is better when we read secrets multiple
times per request.
    
This shows a significant performance improvement when running a
benchmarking script (https://git.io/fpQnW):
    
Before:
    
$ python2.7 benchmark-secrets.py
Time elapsed to create 100 spans = 0.9014 seconds (90.14μs / per)
$ python3.4 benchmark-secrets.py
Time elapsed to create 100 spans = 0.9194 seconds (91.94μs / per)
    
After:
    
$ python2.7 benchmark-secrets.py
Time elapsed to create 100 spans = 0.0352 seconds (3.52μs / per)
$ python3.4 benchmark-secrets.py
Time elapsed to create 100 spans = 0.0381 seconds (3.81μs / per)

Note: this is slightly different from the version proposed for GraphQL. 
It uses `@cached_property` to make the initial loading truly lazy so 
we don't force-fetch the file once per request even if we don't use it.

:eyeglasses: @jesseyli @bsimpson63 @pacejackson