@spladug 2018-12-11
Note: we'd get another nice boost if we cached base64 decoding.