@spladug 2018-12-11
Ran some more tests to ensure that single-secret-per-request perf wasn't worse and we did indeed see a reduction when doing multiple secrets:

```
Before:

$ python2.7 benchmark-secrets.py
Time elapsed to create 250 spans accessing 1 secrets = 0.0282 seconds (112.68μs / per)
Time elapsed to create 250 spans accessing 10 secrets = 0.2159 seconds (863.62μs / per)
Time elapsed to create 250 spans accessing 100 secrets = 2.1297 seconds (8518.98μs / per)
$ python3.4 benchmark-secrets.py
Time elapsed to create 250 spans accessing 1 secrets = 0.0282 seconds (112.94μs / per)
Time elapsed to create 250 spans accessing 10 secrets = 0.2341 seconds (936.49μs / per)
Time elapsed to create 250 spans accessing 100 secrets = 2.1499 seconds (8599.43μs / per)

After:

$ python2.7 benchmark-secrets.py
Time elapsed to create 250 spans accessing 1 secrets = 0.0297 seconds (118.87μs / per)
Time elapsed to create 250 spans accessing 10 secrets = 0.0342 seconds (136.90μs / per)
Time elapsed to create 250 spans accessing 100 secrets = 0.0940 seconds (376.12μs / per)
$ python3.4 benchmark-secrets.py
Time elapsed to create 250 spans accessing 1 secrets = 0.0263 seconds (105.28μs / per)
Time elapsed to create 250 spans accessing 10 secrets = 0.0349 seconds (139.54μs / per)
Time elapsed to create 250 spans accessing 100 secrets = 0.0979 seconds (391.45μs / per)
```